# Handle final elements in SpanHelpers.Contains for byte and char vectorized