@gfoidl 2022-04-03
> It seems we're missing benchmarks for this? (If so can we add yours?)

Sure 😃 https://github.com/dotnet/performance/pull/2347