@AndyAyersMS 2022-04-27
Did we expect any perf regressions from this? Seems like it might be related to https://github.com/dotnet/perf-autofiling-issues/issues/4884.