@gfoidl 2022-04-30
> Did we expect any perf regressions from this? 

No regression is expected, rather it should be an improvement.

When I check the [benchmark-code](https://github.com/dotnet/performance/blob/5dc6276cb84865cb9cd5e24f6c8f862289fa2f98/src/benchmarks/micro/libraries/System.Text.Encoding/Perf.Encoding.cs#L36-L40) `SpanHelpers.Contains` isn't hit?

What is a proper way to investigate this regression?